ver 5.11: refresh version -2

+++ xchange-coinfloor/src/main/java/org/knowm/xchange/coinfloor/service/CoinfloorTradeHistoryParams.java
@@ -1,5 +1,7 @@
 package org.knowm.xchange.coinfloor.service;
 
+import java.util.Collection;
+import java.util.Collections;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.service.trade.params.TradeHistoryParamCurrencyPair;
 import org.knowm.xchange.service.trade.params.TradeHistoryParamMultiCurrencyPair;
@@ -7,82 +9,79 @@ import org.knowm.xchange.service.trade.params.TradeHistoryParamOffset;
 import org.knowm.xchange.service.trade.params.TradeHistoryParamPaging;
 import org.knowm.xchange.service.trade.params.TradeHistoryParamsSorted;
 
-import java.util.Collection;
-import java.util.Collections;
-
 public class CoinfloorTradeHistoryParams
-		implements TradeHistoryParamMultiCurrencyPair,
-		TradeHistoryParamCurrencyPair,
-		TradeHistoryParamsSorted,
-		TradeHistoryParamOffset,
-		TradeHistoryParamPaging {
-	private Collection<CurrencyPair> pairs = Collections.emptySet();
-	private CurrencyPair pair = null;
-	private Order order = null;
-	private Long offset = null;
-	private Integer pageLength = null;
+    implements TradeHistoryParamMultiCurrencyPair,
+        TradeHistoryParamCurrencyPair,
+        TradeHistoryParamsSorted,
+        TradeHistoryParamOffset,
+        TradeHistoryParamPaging {
+  private Collection<CurrencyPair> pairs = Collections.emptySet();
+  private CurrencyPair pair = null;
+  private Order order = null;
+  private Long offset = null;
+  private Integer pageLength = null;
 
-	@Override
-	public Collection<CurrencyPair> getCurrencyPairs() {
-		return pairs;
-	}
+  @Override
+  public Collection<CurrencyPair> getCurrencyPairs() {
+    return pairs;
+  }
 
-	@Override
-	public void setCurrencyPairs(Collection<CurrencyPair> value) {
-		pairs = value;
-	}
+  @Override
+  public void setCurrencyPairs(Collection<CurrencyPair> value) {
+    pairs = value;
+  }
 
-	@Override
-	public CurrencyPair getCurrencyPair() {
-		return pair;
-	}
+  @Override
+  public CurrencyPair getCurrencyPair() {
+    return pair;
+  }
 
-	@Override
-	public void setCurrencyPair(CurrencyPair value) {
-		pair = value;
-	}
+  @Override
+  public void setCurrencyPair(CurrencyPair value) {
+    pair = value;
+  }
 
-	@Override
-	public Order getOrder() {
-		return order;
-	}
+  @Override
+  public Order getOrder() {
+    return order;
+  }
 
-	@Override
-	public void setOrder(Order order) {
-		this.order = order;
-	}
+  @Override
+  public void setOrder(Order order) {
+    this.order = order;
+  }
 
-	@Override
-	public Long getOffset() {
-		return offset;
-	}
+  @Override
+  public Long getOffset() {
+    return offset;
+  }
 
-	@Override
-	public void setOffset(Long offset) {
-		this.offset = offset;
-	}
+  @Override
+  public void setOffset(Long offset) {
+    this.offset = offset;
+  }
 
-	@Override
-	public Integer getPageLength() {
-		return pageLength;
-	}
+  @Override
+  public Integer getPageLength() {
+    return pageLength;
+  }
 
-	@Override
-	public void setPageLength(Integer pageLength) {
-		this.pageLength = pageLength;
-	}
+  @Override
+  public void setPageLength(Integer pageLength) {
+    this.pageLength = pageLength;
+  }
 
-	@Override
-	public Integer getPageNumber() {
-		return (offset == null || pageLength == null) ? null : offset.intValue() / pageLength;
-	}
+  @Override
+  public Integer getPageNumber() {
+    return (offset == null || pageLength == null) ? null : offset.intValue() / pageLength;
+  }
 
-	@Override
-	public void setPageNumber(Integer pageNumber) {
-		if (pageNumber == null) {
-			setOffset(null);
-		} else if (pageLength != null) {
-			this.offset = Long.valueOf((long) pageNumber * pageLength);
-		}
-	}
+  @Override
+  public void setPageNumber(Integer pageNumber) {
+    if (pageNumber == null) {
+      setOffset(null);
+    } else if (pageLength != null) {
+      this.offset = Long.valueOf(pageNumber * pageLength);
+    }
+  }
 }
